cleanup / reorg and write todos
